pyproject.toml
[project]
name = "crud-funcionarios"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi==0.135.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary==2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy==2.0.48",
    "uvicorn==0.42.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.2",
]
